requirements.txt
requests
python-dotenv
pandas
numpy
matplotlib
seaborn
cartopy